+++ R/assert_webmorph.R
@@ -1,9 +1,10 @@
 #' Check for webmorph_list
 #'
-#' Check if an object is a list of class webmorph_list. If class webmorph_stim, wrap in webmorph_list. If a properly structured list without the right class, add the right class.
+#' Check if an object is a list of class webmorph_list. If class webmorph_stim, wrap in webmorph_list. If a properly structured list without the right class, add the right class. If the img is not a magick_image or the pointer is dead, reloads from the imgpath.
 #'
 #' @param x The object
-#' @param arg The name of the argument
+#'
+#' @keywords internal
 #'
 #' @return A webmorph_list
 #'
@@ -31,34 +32,36 @@ assert_webmorph <- function(x) {
   # convert webmorph_stim to webmorph_list
   if ("webmorph_stim" %in% class(x)) {
     #convert to webmorph_list
-    stim_list <- list(x)
-    class(stim_list) <- "webmorph_list"
+    stimlist <- list(x)
+    class(stimlist) <- "webmorph_list"
   } else if ("webmorph_list" %in% class(x)) {
-    stim_list <- x
+    stimlist <- x
   } else {
     arg <- match.call()$x
     stop(arg, " needs to be a webmorph_list")
   }
 
   # add names
-  if (is.null(names(stim_list))) {
-    i <- sapply(stim_list, `[[`, "imgpath")
-    t <- sapply(stim_list, `[[`, "tempath")
+  if (is.null(names(stimlist))) {
+    i <- sapply(stimlist, `[[`, "imgpath")
+    t <- sapply(stimlist, `[[`, "tempath")
     nm <- mapply(function(i, t) { i %||% t }, i, t)
-    names(stim_list) <- unique_names(nm)
+    names(stimlist) <- unique_names(nm)
   }
 
   # check images are available and reload if not
-  for (i in seq_along(stim_list)) {
-    img <- stim_list[[i]]$img
+  for (i in seq_along(stimlist)) {
+    img <- stimlist[[i]]$img
     if (!is.null(img)) {
-      tryCatch({
-        magick:::assert_image(img)
+      stimlist[[i]]$img <- tryCatch({
+        magick::image_info(img) # throws an error if img not valid
+        # rather use magick:::assert_image, but ::: warning
+        img
       }, error = function(e) {
-        stim_list[[i]]$img <- magick::image_read(stim_list[[i]]$imgpath)
+        magick::image_read(stimlist[[i]]$imgpath)
       })
     }
   }
 
-  stim_list
+  stimlist
 }