+++ tests/testthat/test-assert_webmorph.R
@@ -1,11 +1,11 @@
 path <- system.file("extdata/test", package = "webmorph")
-stim_list <- read_stim(path)
-stim <- stim_list[[1]]
+stimlist <- read_stim(path)
+stim <- stimlist[[1]]
 
 test_that("works", {
   # valid webmorph_list
-  x <- assert_webmorph(stim_list)
-  expect_equal(x, stim_list)
+  x <- assert_webmorph(stimlist)
+  expect_equal(x, stimlist)
 
   # valid webmorph_stim
   x <- assert_webmorph(stim)
@@ -13,12 +13,12 @@ test_that("works", {
   expect_equal(names(x), "f_multi")
 
   # valid webmorph_list with no class
-  x <- assert_webmorph(unclass(stim_list))
-  expect_equal(x, stim_list)
+  x <- assert_webmorph(unclass(stimlist))
+  expect_equal(x, stimlist)
 
   # valid webmorph_list with no names
-  x <- assert_webmorph(unname(stim_list))
-  expect_equal(x, stim_list)
+  x <- assert_webmorph(unname(stimlist))
+  expect_equal(x, stimlist)
 
   # valid webmorph_stim with no class
   x <- assert_webmorph(unclass(stim))
@@ -31,4 +31,10 @@ test_that("errors", {
   expect_error(assert_webmorph(c()))
   expect_error(assert_webmorph(list()))
   expect_error(assert_webmorph(list("a")))
+})
+
+test_that("reload img", {
+  stim$img <- NULL
+  stim2 <- webmorph:::assert_webmorph(stim)
+  expect_equal(class(stim2[[1]]$img), "magick-image")
 })