+++ NEWS.md
@@ -1,3 +1,8 @@
+# webmorph 0.0.0.9001
+
+* Cleaned up functions and examples
+* Added `align()` function
+
 # webmorph 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.